MONTGOMERY - Bug 28966: Prefetch patron data for holds queue viewer
We have a partner processing thousands of holds per day, per branch. At this number of holds, the script loads very slow if it manages to load at all. The primary slowdown is the individual fetch of the patron for each hold to be displayed. If we prefetch those patrons, the entire script should load much faster.

Test Plan:
1) Create as many holds as you can ( up to a few thousand if you can )
2) Run the holds queue viewer, note the load time
3) Apply this patch
4) Restart all the things!
5) Reload the holds queue viewer, load time should be improved
6) Note the HTML generated looks the same