BWS-PKG - Bug 29474: Return too_soon before checking renewals
This patch allows us to avoid checking reserves when the issue is
too_soon and we are running the cron

Code is covered by existing tests

To test:
1 - prove -v t/db_dependent/Circulation.t

Bug 29474: Decouple renewal checks of auto-renewed issues from non-autorenewed

The too_soon check does not relate at all to auto-renewed issues and
should be moved outside the _CanBookBeAutoRenewed() function.

To test:
   1) prove t/db_dependent/Circulation.t

Bug 29474: Remove one layer of indendation by adding if check in the begginning

If we simply return "no" immediately from the function when the
checkout is not an autorenewed checkout we can drop one layer of
indendation and the code becomes much easier to read.

To test:
 1) prove t/db_dependent/Circulation.t

Bug 29474: Pass rules/branchcode to subroutines, further simplify logic

If we are going to decouple 'too_soon' from auto_renewals we should only call GetSoonestDate once

Additionally, we can fetch all needed rules at once and pass to subroutines as needed

Rather than check in the sub, let's not call _CanBookBeAutoRenewed if we aren't set for auto_renew

Bug 29474: Small efficiency improvement