CUYAHOGA - Bug 34160: Hyperlink item barcode to the item more details editor from the holds queue viewer
Sometimes librarians fail to find an item on the shelf for a hold and need to mark the item as lost. In these cases it would be helpful to link the barcode to the item's "more details" form in the same way we do for the standard details item barcode hyperlink.

Test Plan:
1) View the holds queue, note the item barcode is not a hyperlink
2) Apply this patch
3) Reload the page, note the item barcode is a hyperlink to the more
   details page for that item!

Signed-off-by: Phil Ringnalda <phil@chetcolibrary.org>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>

Bug 34160: Fix filters in URL from html to uri

Following the advice of the mighty QA test tools :)

Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>